renamed gmap app to lib, and gmap view to get_map
