+++ the_game_bazaar/settings.py
@@ -127,8 +127,8 @@ INSTALLED_APPS = (
     'django.contrib.staticfiles',
     'django_extensions',
     'the_game_bazaar',
+    'lib',
     'game',
-    'gmap',
     'editor',
     'jasmine_testing',
     # Uncomment the next line to enable the admin:
