+++ the_game_bazaar/views.py
@@ -7,7 +7,7 @@ from django.contrib.auth import logout
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import login as auth_login
-from gmap.models import Map
+from lib.models import Map
 from game.models import Game
